feat: implemented self created bike deletion for USER role and any bike deletion for ADMIN role

+++ src/controllers/bikes-controller/mutations/delete-bike.ts
@@ -1,5 +1,6 @@
 import { RequestHandler } from 'express';
-import ErrorService, { ServerSetupError } from 'services/error-service';
+import ErrorService, { ServerSetupError, ForbiddenError } from 'services/error-service';
+import UserModel from 'models/user-model';
 import { BikeViewModel } from '../types';
 import BikeModel from '../model';
 
@@ -13,7 +14,12 @@ BikeViewModel | ErrorResponse,
 
   try {
     if (id === undefined) throw new ServerSetupError();
+    if (req.authData === undefined) throw new ServerSetupError();
+    const user = await UserModel.getUserByEmail(req.authData.email);
     const bike = await BikeModel.getBike(id);
+
+    if (user.role !== 'ADMIN' && user.id !== bike.owner.id) throw new ForbiddenError();
+
     await BikeModel.deleteBike(id);
 
     res.status(200).json(bike);
